exchanges: reverts to old style in GetFeeByType, reverts some code I accidently deleted. Plus things and other. XD
