+++ exchanges/gemini/gemini_wrapper.go
@@ -663,8 +663,7 @@ func (g *Gemini) GetFeeByType(ctx context.Context, feeBuilder *exchange.FeeBuild
 	if feeBuilder == nil {
 		return 0, fmt.Errorf("%T %w", feeBuilder, common.ErrNilPointer)
 	}
-	_, err := g.GetCredentials(ctx)
-	if (err != nil || g.SkipAuthCheck) && // Todo check connection status
+	if (!g.AreCredentialsValid(ctx) || g.SkipAuthCheck) && // Todo check connection status
 		feeBuilder.FeeType == exchange.CryptocurrencyTradeFee {
 		feeBuilder.FeeType = exchange.OfflineTradeFee
 	}
