+++ exchanges/poloniex/poloniex_wrapper.go
@@ -806,8 +806,7 @@ func (p *Poloniex) GetFeeByType(ctx context.Context, feeBuilder *exchange.FeeBui
 	if feeBuilder == nil {
 		return 0, fmt.Errorf("%T %w", feeBuilder, common.ErrNilPointer)
 	}
-	_, err := p.GetCredentials(ctx)
-	if (err != nil || p.SkipAuthCheck) && // Todo check connection status
+	if (!p.AreCredentialsValid(ctx) || p.SkipAuthCheck) && // Todo check connection status
 		feeBuilder.FeeType == exchange.CryptocurrencyTradeFee {
 		feeBuilder.FeeType = exchange.OfflineTradeFee
 	}
