+++ exchanges/zb/zb_wrapper.go
@@ -693,8 +693,7 @@ func (z *ZB) GetFeeByType(ctx context.Context, feeBuilder *exchange.FeeBuilder) 
 	if feeBuilder == nil {
 		return 0, fmt.Errorf("%T %w", feeBuilder, common.ErrNilPointer)
 	}
-	_, err := z.GetCredentials(ctx)
-	if (err != nil || z.SkipAuthCheck) && // Todo check connection status
+	if (!z.AreCredentialsValid(ctx) || z.SkipAuthCheck) && // Todo check connection status
 		feeBuilder.FeeType == exchange.CryptocurrencyTradeFee {
 		feeBuilder.FeeType = exchange.OfflineTradeFee
 	}
